Fix code scanning alert #17: Multiplication result converted to larger type

+++ src/ai/formula/function_table.cpp
@@ -278,7 +278,7 @@ DEFINE_FAI_FUNCTION(calculate_map_ownership, 2, 5)
 	std::vector< std::vector<int>> scores( number_of_teams );
 
 	for( std::size_t i = 0; i< number_of_teams; ++i)
-		scores[i].resize(w*h);
+		scores[i].resize(static_cast<std::size_t>(w) * h);
 
 	/* // TODO: Do we need this?
 	for(unit_map::const_iterator i = resources::gameboard->units().begin(); i != resources::gameboard->units().end(); ++i) {
